connectmeapp_transfer_cleaned/functions/requirements.txt
firebase-admin>=6.0.0
firebase-functions>=0.4.0
Flask>=3.0.0
flask-cors>=4.0.0
functions-framework>=3.0.0
geohash2>=1.1
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
stripe>=5.0.0
gunicorn>=21.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
requests>=2.31.0
